# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/assinatura": {
            "get": {
                "tags": [
                    "Assinaturas"
                ],
                "summary": "Buscar assinaturas",
                "description": "Todas as assinaturas",
                "operationId": "096a6fdf2f01b6e5583da8fcfc593e75",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "data_inicio": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "data_termino": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "plano_id": {
                                                        "type": "integer"
                                                    },
                                                    "cliente_id": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "errcode": 1,
                                        "errmsg": "ok",
                                        "data": []
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Assinaturas"
                ],
                "summary": "Assinatura",
                "operationId": "9e205cad62c739d10e5d674d53a2774b",
                "parameters": [
                    {
                        "name": "plano",
                        "in": "header",
                        "description": "id do plano",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "cliente",
                        "in": "header",
                        "description": "Id do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "vigencia",
                        "in": "header",
                        "description": "Dias de vigência do plano",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Assinatura registrada com sucesso"
                    },
                    "422": {
                        "description": "Erro na validação"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/assinatura/{id}": {
            "get": {
                "tags": [
                    "Assinaturas"
                ],
                "summary": "Buscar um assinatura especifica",
                "operationId": "f56275d8ef8b60599adb232649937868",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id da assinatura",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data_inicio": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "data_termino": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "plano_id": {
                                            "type": "integer"
                                        },
                                        "cliente_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Assinaturas"
                ],
                "summary": "Excluir uma assinatura",
                "operationId": "f6d39e50ec2e70c44d95f85a5f0d4607",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id da assinatura",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Assinatura exluida com sucesso"
                    },
                    "422": {
                        "description": "Erro na validação"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Assinaturas"
                ],
                "summary": "Atualizar Assinatura",
                "operationId": "099c979d12baded32ed7ff3a8e25aa0a",
                "parameters": [
                    {
                        "name": "data_termino",
                        "in": "header",
                        "description": "No data de termino",
                        "required": true,
                        "schema": {
                            "type": "datetime"
                        }
                    },
                    {
                        "name": "status",
                        "in": "header",
                        "description": "Novo status ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id da assinatura",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Assinatura atualizada com sucesso"
                    },
                    "422": {
                        "description": "Erro na validação"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/cliente/{id}": {
            "get": {
                "tags": [
                    "Assinaturas"
                ],
                "summary": "Buscar as assinaturas do cliente",
                "operationId": "0a0587ec20157cff55f61d19ed91fc89",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id da cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "data_inicio": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "data_termino": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "plano_id": {
                                                        "type": "integer"
                                                    },
                                                    "cliente_id": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "errcode": 1,
                                        "errmsg": "ok",
                                        "data": []
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}